+++ biz/webui/vite-dev/vite.config.js
@@ -1,5 +1,5 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
 /**
  * @returns {import('vite').PluginOption}
@@ -7,36 +7,52 @@ import react from '@vitejs/plugin-react';
 function wrapperCJS() {
   let ignore = [
     '\'./files-dialog\''
-  ];
+  ]
   return {
     name: 'wrapperCJS',
     async transform(code, id) {
       if (id.split('?')[0].endsWith('.js')) {
         // wrapper cjs
-        let matches = code.match(/require\([^)]+\)/g) || [];
+        let matches = code.match(/require\([^)]+\)/g) || []
 
         if (matches.length == 0 && !code.includes('exports')) {
-          return code;
+          return code
         }
 
-        let arr = [];
+        let arr = []
         for (let i = 0; i < matches.length; i++) {
-          const match = matches[i];
-          const id = match.replace('require(', '').replace(')', '');
+          const match = matches[i]
+          const id = match.replace('require(', '').replace(')', '')
           if (ignore.includes(id)) {
-            continue;
+            continue
           }
-          arr.push({ name: `__import_key_${i}__`, id });
+          arr.push({ name: `__import_key_${i}__`, id })
         }
 
         return `
 
-${arr.map(o => `import * as ${o.name} from ${o.id}`).join('\n')}
+${arr.map(o => `import * as ${o.name} from ${addCSSInline(o.id)}`).join('\n')}
+
+function __$styleInject_43268484ff8556ef3adf6e1d73e072de(css) {
+  if (!css) return;
+
+  if (typeof window == 'undefined') return;
+  var style = document.createElement('style');
+  style.setAttribute('media', 'screen');
+
+  style.innerHTML = css;
+  document.head.appendChild(style);
+  return css;
+}
 
 function require(id){
-  return {
+  let mod = {
     ${arr.map(o => `${o.id}: ${o.name}.default || ${o.name},`).join('\n')}
-  }[id]
+  }[id];
+
+  id.endsWith('.css') && __$styleInject_43268484ff8556ef3adf6e1d73e072de(mod);
+
+  return mod
 }
 
 const module={
@@ -59,10 +75,17 @@ if(exports.__esModule && !exports.default){
 }
 
 ${code.includes('export ') ? '' : 'export default exports'}
-          `;
+          `
       }
     }
-  };
+  }
+}
+
+function addCSSInline(id) {
+  if (id.endsWith(`.css'`)) {
+    id = id.replace(/css'$/, `css?inline'`)
+  }
+  return id
 }
 
 /**
@@ -73,10 +96,10 @@ function fileloader(ext) {
     name: 'fileloader:' + ext,
     async transform(code, id) {
       if (id.endsWith('.md')) {
-        return 'export default ' + JSON.stringify(code);
+        return 'export default ' + JSON.stringify(code)
       }
     }
-  };
+  }
 }
 
 /**
@@ -85,11 +108,11 @@ function fileloader(ext) {
 function fixedModule() {
   return {
     transform(code, id) {
-      if(id.includes('bootstrap_dist_js_bootstrap__js')){
-        return code + '\n exports = {}';
+      if (id.includes('bootstrap_dist_js_bootstrap__js')) {
+        return code + '\n exports = {}'
       }
     }
-  };
+  }
 }
 
 // https://vitejs.dev/config/